Refactored Loading into a component for improved reusability. States are now screens.
