+++ app/src/main/java/com/example/easyepubtransfer/MainActivity.kt
@@ -11,8 +11,7 @@ import androidx.compose.material3.Scaffold
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
-import com.example.easyepubtransfer.ui.states.DefaultView
-import com.example.easyepubtransfer.ui.states.LoadingView
+import com.example.easyepubtransfer.ui.screens.DefaultView
 import com.example.easyepubtransfer.ui.theme.EasyEPUBTransferTheme
 
 class MainActivity : ComponentActivity() {
