+++ app/src/main/java/com/example/easyepubtransfer/ui/screens/DefaultView.kt
@@ -1,4 +1,4 @@
-package com.example.easyepubtransfer.ui.states
+package com.example.easyepubtransfer.ui.screens
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
